+++ R/metrics-sample.R
@@ -92,7 +92,7 @@ bias_sample <- function(observed, predicted) {
 #' the number of data points and N (number of columns) the number of Monte
 #' Carlo samples. Alternatively, `predicted` can just be a vector of size n.
 #' @return vector with the scoring values
-#' @seealso [ae_median_quantile()], [abs_error()]
+#' @seealso [ae_median_quantile()]
 #' @importFrom stats median
 #' @examples
 #' observed <- rnorm(30, mean = 1:30)
@@ -127,7 +127,6 @@ ae_median_sample <- function(observed, predicted) {
 #' the number of data points and N (number of columns) the number of Monte
 #' Carlo samples. Alternatively, `predicted` can just be a vector of size n.
 #' @return vector with the scoring values
-#' @seealso [squared_error()]
 #' @examples
 #' observed <- rnorm(30, mean = 1:30)
 #' predicted_values <- matrix(rnorm(30, mean = 1:30))
@@ -275,45 +274,10 @@ crps_sample <- function(observed, predicted, ...) {
 #' predicted <- replicate(200, rpois(n = 30, lambda = 1:30))
 #' mad_sample(predicted = predicted)
 #' @keywords metric
-
 mad_sample <- function(observed = NULL, predicted, ...) {
 
   assert_input_sample(rep(NA_real_, nrow(predicted)), predicted)
 
   sharpness <- apply(predicted, MARGIN = 1, mad, ...)
   return(sharpness)
-}
-
-
-#' @title Interval Coverage
-#' @description To compute coverage for sample-based forecasts,
-#' predictive samples are converted first into predictive quantiles using the
-#' sample quantiles.
-#' @importFrom checkmate assert_number
-#' @rdname interval_coverage
-#' @export
-#' @examples
-#' interval_coverage_sample(observed, predicted)
-interval_coverage_sample <- function(observed, predicted, range = 50) {
-  assert_input_sample(observed, predicted)
-  assert_number(range)
-  necessary_quantiles <- c((100 - range) / 2, 100 - (100 - range) / 2) / 100
-
-  # this could be its own function, sample_to_quantile.numeric
-  # ==========================================================
-  n <- length(observed)
-  N <- length(predicted) / n
-  dt <- data.table(
-    observed = rep(observed, each = N),
-    predicted = as.vector(t(predicted))
-  )
-  quantile_dt <- sample_to_quantile(dt, necessary_quantiles)
-  # ==========================================================
-
-  # this could call interval_coverage_quantile instead
-  # ==========================================================
-  interval_dt <- quantile_to_interval(quantile_dt, format = "wide")
-  interval_dt[, coverage := (observed >= lower) & (observed <= upper)]
-  # ==========================================================
-  return(interval_dt$coverage)
 }