+++ R/score.R
@@ -32,7 +32,9 @@
 #' @importFrom stats na.omit
 #' @examples
 #' library(magrittr) # pipe operator
-#' data.table::setDTthreads(1) # only needed to avoid issues on CRAN
+#' \dontshow{
+#'   data.table::setDTthreads(2) # restricts number of cores used on CRAN
+#' }
 #'
 #' validated <- as_forecast(example_quantile)
 #' score(validated) %>%
@@ -76,6 +78,7 @@ score.default <- function(data, ...) {
 }
 
 #' @importFrom stats na.omit
+#' @importFrom data.table setattr
 #' @rdname score
 #' @export
 score.forecast_binary <- function(data, metrics = rules_binary(), ...) {
@@ -83,20 +86,19 @@ score.forecast_binary <- function(data, metrics = rules_binary(), ...) {
   data <- na.omit(data)
   metrics <- validate_metrics(metrics)
 
-  data <- apply_rules(
+  scores <- apply_rules(
     data, metrics,
     data$observed, data$predicted, ...
   )
 
-  setattr(data, "metric_names", names(metrics))
-
-  return(data[])
-
+  scores <- as_scores(scores, score_names = names(metrics))
+  return(scores[])
 }
 
 
 #' @importFrom Metrics se ae ape
 #' @importFrom stats na.omit
+#' @importFrom data.table setattr
 #' @rdname score
 #' @export
 score.forecast_point <- function(data, metrics = rules_point(), ...) {
@@ -104,23 +106,23 @@ score.forecast_point <- function(data, metrics = rules_point(), ...) {
   data <- na.omit(data)
   metrics <- validate_metrics(metrics)
 
-  data <- apply_rules(
+  scores <- apply_rules(
     data, metrics,
     data$observed, data$predicted, ...
   )
 
-  setattr(data, "metric_names", names(metrics))
-
-  return(data[])
+  scores <- as_scores(scores, score_names = names(metrics))
+  return(scores[])
 }
 
 #' @importFrom stats na.omit
+#' @importFrom data.table setattr
 #' @rdname score
 #' @export
 score.forecast_sample <- function(data, metrics = rules_sample(), ...) {
   data <- validate_forecast(data)
   data <- na.omit(data)
-  forecast_unit <- attr(data, "forecast_unit")
+  forecast_unit <- get_forecast_unit(data)
   metrics <- validate_metrics(metrics)
 
   # transpose the forecasts that belong to the same forecast unit
@@ -145,20 +147,20 @@ score.forecast_sample <- function(data, metrics = rules_sample(), ...) {
     )
     return(data)
   })
-  data <- rbindlist(split_result)
-  setattr(data, "metric_names", names(metrics))
-
-  return(data[])
+  scores <- rbindlist(split_result)
+  scores <- as_scores(scores, score_names = names(metrics))
+  return(scores[])
 }
 
+
 #' @importFrom stats na.omit
-#' @importFrom data.table `:=` as.data.table rbindlist %like%
+#' @importFrom data.table `:=` as.data.table rbindlist %like% setattr
 #' @rdname score
 #' @export
 score.forecast_quantile <- function(data, metrics = rules_quantile(), ...) {
   data <- validate_forecast(data)
   data <- na.omit(data)
-  forecast_unit <- attr(data, "forecast_unit")
+  forecast_unit <- get_forecast_unit(data)
   metrics <- validate_metrics(metrics)
 
   # transpose the forecasts that belong to the same forecast unit
@@ -190,11 +192,13 @@ score.forecast_quantile <- function(data, metrics = rules_quantile(), ...) {
     )
     return(data)
   })
+  scores <- rbindlist(split_result)
 
-  data <- rbindlist(split_result)
-  setattr(data, "metric_names", names(metrics))
+  # this can have existing scores, e.g. from `add_coverage()`
+  existing_scores <- get_score_names(data)
+  scores <- as_scores(scores, score_names = c(existing_scores, names(metrics)))
 
-  return(data[])
+  return(scores[])
 }
 
 
@@ -218,4 +222,59 @@ apply_rules <- function(data, metrics, ...) {
     eval(expr)
   }, data, ...)
   return(data)
+}
+
+
+#' Construct An Object Of Class `scores`
+#' @description This function creates an object of class `scores` based on a
+#' data.table or similar.
+#' @param scores A data.table or similar with scores as produced by [score()]
+#' @param score_names A character vector with the names of the scores
+#' (i.e. the names of the scoring rules used for scoring)
+#' @keywords internal
+#' @examples
+#' \dontrun{
+#' df <- data.frame(
+#'   model = "A",
+#'   wis = "0.1"
+#' )
+#' new_scores(df, "wis")
+#' }
+new_scores <- function(scores, score_names) {
+  scores <- as.data.table(scores)
+  class(scores) <- c("scores", class(scores))
+  setattr(scores, "score_names", score_names)
+  return(scores[])
+}
+
+
+#' Create An Object Of Class `scores` From Data
+#' @description This convenience function wraps [new_scores()] and validates
+#' the `scores` object.
+#' @inheritParams new_scores
+#' @returns Returns an object of class 1scores`
+#' @importFrom checkmate assert_data_frame
+#' @keywords internal
+as_scores <- function(scores, score_names) {
+  assert_data_frame(scores)
+  scores <- new_scores(scores, score_names)
+  validate_scores(scores)
+  return(scores[])
+}
+
+
+#' Validate An Object Of Class `scores`
+#' @description This function validates an object of class `scores`, checking
+#' that it has the correct class and that it has a `score_names` attribute.
+#' @inheritParams new_scores
+#' @returns Returns `NULL` invisibly
+#' @importFrom checkmate assert_class assert_data_frame
+#' @keywords internal
+validate_scores <- function(scores) {
+  assert_data_frame(scores)
+  assert_class(scores, "scores")
+  # error if no score_names exists +
+  # throw warning if any of the score_names is not in the data
+  get_score_names(scores, error = TRUE)
+  return(invisible(NULL))
 }