+++ tests/testthat/test-metrics-quantile.R
@@ -585,37 +585,54 @@ test_that("wis is the sum of overprediction, underprediction, dispersion", {
 
 
 # ============================================================================ #
-# `interval_coverage_quantile` =============================================== #
+# `interval_coverage` =============================================== #
 # ============================================================================ #
-test_that("interval_coverage_quantile works", {
+test_that("interval_coverage works", {
   expect_equal(
-    interval_coverage_quantile(observed, predicted, quantile, range = 50),
+    interval_coverage(observed, predicted, quantile, range = 50),
     c(TRUE, FALSE, FALSE)
   )
 })
 
-test_that("interval_coverage_quantile rejects wrong inputs", {
+test_that("interval_coverage rejects wrong inputs", {
   expect_error(
-    interval_coverage_quantile(observed, predicted, quantile, range = c(50, 0)),
+    interval_coverage(observed, predicted, quantile, range = c(50, 0)),
     "Assertion on 'range' failed: Must have length 1."
+  )
+})
+
+test_that("interval_coverage_quantile throws a warning when a required quantile is not available", {
+  dropped_quantile_pred <- predicted[, -4]
+  dropped_quantiles <- quantile[-4]
+  expect_warning(
+    interval_coverage(
+      observed, dropped_quantile_pred, dropped_quantiles, range = 50
+    ),
+    "To compute the interval coverage for a range of 50%, the quantiles `0.25, 0.75` are required. Returning `NA`"
   )
 })
 
 
 # ============================================================================ #
-# `interval_coverage_dev_quantile` ===================================== #
+# `interval_coverage_deviation` ===================================== #
 # ============================================================================ #
-test_that("interval_coverage_dev_quantile works", {
+test_that("interval_coverage_deviation works", {
   existing_ranges <- unique(get_range_from_quantile(quantile))
   expect_equal(existing_ranges, c(80, 50, 0))
 
-  cov_50 <- interval_coverage_quantile(observed, predicted, quantile, range = c(50))
-  cov_80 <- interval_coverage_quantile(observed, predicted, quantile, range = c(80))
+  cov_50 <- interval_coverage(observed, predicted, quantile, range = c(50))
+  cov_80 <- interval_coverage(observed, predicted, quantile, range = c(80))
   manual <- 0.5 * (cov_50 - 0.5) + 0.5 * (cov_80 - 0.8)
 
   expect_equal(
-    interval_coverage_dev_quantile(observed, predicted, quantile),
+    interval_coverage_deviation(observed, predicted, quantile),
     manual
+  )
+  expect_warning(
+    interval_coverage_deviation(
+      observed, predicted, c(quantile[-4], 0.76)
+    ),
+    "To compute inteval coverage deviation, all quantiles must form central symmetric prediction intervals. Missing quantiles: 0.24, 0.75. Returning `NA`."
   )
 })